+++ lerna-management/src/main/scala/lerna/management/stats/MetricsImpl.scala
@@ -4,8 +4,10 @@ import akka.actor.{ ActorRef, ActorSystem }
 import akka.pattern.ask
 import akka.util.Timeout
 import com.typesafe.config.Config
+import kamon.Kamon
 import lerna.util.lang.Equals._
-import kamon.metric.{ MetricDistribution, MetricValue, PeriodSnapshot }
+import kamon.metric.{ Distribution, MetricSnapshot, PeriodSnapshot }
+import kamon.tag.TagSet
 import lerna.management.stats.MetricsActor.{ GetMetrics, UpdateMetrics }
 import lerna.util.tenant.Tenant
 
@@ -13,7 +15,7 @@ import scala.concurrent.Future
 import scala.concurrent.duration._
 import scala.math.BigDecimal.RoundingMode
 
-private[stats] final case class SettingItem(key: MetricsKey, name: String, tags: kamon.Tags, nullValue: Option[String])
+private[stats] final case class SettingItem(key: MetricsKey, name: String, tags: TagSet, nullValue: Option[String])
 
 @SuppressWarnings(
   Array(
@@ -22,25 +24,16 @@ private[stats] final case class SettingItem(key: MetricsKey, name: String, tags:
 )
 private[stats] class MetricsImpl(system: ActorSystem, tenants: Set[Tenant]) extends Metrics {
 
+  private val kamonModuleName            = "lerna-app-library-metrics"
   private var settings: Set[SettingItem] = readSettings(system.settings.config)
   private val actor: ActorRef            = system.actorOf(MetricsActor.props())
 
   implicit lazy val timeout: Timeout = Timeout(3.seconds)
 
+  Kamon.registerModule(kamonModuleName, this)
+
   override def getMetrics(key: MetricsKey): Future[Option[MetricsValue]] = {
     (actor ? GetMetrics(key)).mapTo[Option[MetricsValue]]
-  }
-
-  private object SettingExistsMetricValue {
-    def unapply(arg: MetricValue): Option[(SettingItem, MetricValue)] = {
-      settings.find(e => e.name === arg.name && e.tags === arg.tags).map(s => (s, arg))
-    }
-  }
-
-  private object SettingExistsMetricDistribution {
-    def unapply(arg: MetricDistribution): Option[(SettingItem, MetricDistribution)] = {
-      settings.find(e => e.name === arg.name && e.tags === arg.tags).map(s => (s, arg))
-    }
   }
 
   override def reportPeriodSnapshot(snapshot: PeriodSnapshot): Unit = {
@@ -48,20 +41,11 @@ private[stats] class MetricsImpl(system: ActorSystem, tenants: Set[Tenant]) exte
       settings
         .map(e => (e.key, UpdateMetrics(e.key, e.nullValue.map(MetricsValue)))).toMap
 
-    val newMetricValues = (snapshot.metrics.counters ++ snapshot.metrics.gauges)
-      .collect {
-        case SettingExistsMetricValue((item, value)) =>
-          UpdateMetrics(item.key, Option(MetricsValue(value.value.toString)))
-      }
-    val newMetricDistributions = (snapshot.metrics.histograms ++ snapshot.metrics.rangeSamplers)
-      .collect {
-        case SettingExistsMetricDistribution((item, value)) =>
-          val average =
-            if (value.distribution.count === 0L) BigDecimal(0)
-            else (BigDecimal(value.distribution.sum) / value.distribution.count).setScale(0, RoundingMode.HALF_DOWN)
-          UpdateMetrics(item.key, Option(MetricsValue(average.bigDecimal.toPlainString)))
-      }
-    val allMetrics: Seq[UpdateMetrics] = newMetricValues ++ newMetricDistributions
+    val allMetrics: Seq[UpdateMetrics] =
+      snapshot.counters.flatMap(collectUpdateMetricsFromValues) ++
+      snapshot.gauges.flatMap(collectUpdateMetricsFromValues) ++
+      snapshot.histograms.flatMap(collectUpdateMetricsFromDistributions) ++
+      snapshot.rangeSamplers.flatMap(collectUpdateMetricsFromDistributions)
 
     allMetrics
       .foldLeft(emptyMetrics) { (acc, metrics) =>
@@ -72,7 +56,34 @@ private[stats] class MetricsImpl(system: ActorSystem, tenants: Set[Tenant]) exte
       }
   }
 
-  override def start(): Unit = {}
+  private[this] def findSetting(name: String, tags: TagSet): Option[SettingItem] = {
+    settings.find(s => s.name === name && s.tags === tags)
+  }
+
+  private[this] def collectUpdateMetricsFromValues[T](values: MetricSnapshot.Values[T]): Seq[UpdateMetrics] = {
+    values.instruments.flatMap { instrument =>
+      findSetting(values.name, instrument.tags)
+        .map { setting =>
+          UpdateMetrics(setting.key, Option(MetricsValue(instrument.value.toString)))
+        }
+    }
+  }
+
+  private[this] def collectUpdateMetricsFromDistributions(
+      distributions: MetricSnapshot.Distributions,
+  ): Seq[UpdateMetrics] = {
+    def averageOf(distribution: Distribution): BigDecimal = {
+      if (distribution.count === 0L) BigDecimal(0)
+      else (BigDecimal(distribution.sum) / distribution.count).setScale(0, RoundingMode.HALF_DOWN)
+    }
+    distributions.instruments.flatMap { instrument =>
+      findSetting(distributions.name, instrument.tags)
+        .map { setting =>
+          val average = averageOf(instrument.value)
+          UpdateMetrics(setting.key, Option(MetricsValue(average.bigDecimal.toPlainString)))
+        }
+    }
+  }
 
   override def stop(): Unit = {}
 
@@ -89,30 +100,33 @@ private[stats] class MetricsImpl(system: ActorSystem, tenants: Set[Tenant]) exte
     items.flatMap { item =>
       val key    = item.getKey.replaceFirst("^/", "")
       val config = root.getConfig(item.getKey)
-      val tags: Map[String, String] =
-        if (config.hasPath("tags")) {
-          val tagsConfig = config.getConfig("tags")
-          tagsConfig
-            .entrySet().asScala
-            .map(e => (e.getKey, tagsConfig.getString(e.getKey))).toMap
-        } else {
-          Map()
-        }
+      val tagSet: TagSet = {
+        val tags: Map[String, String] =
+          if (config.hasPath("tags")) {
+            val tagsConfig = config.getConfig("tags")
+            tagsConfig
+              .entrySet().asScala
+              .map(e => (e.getKey, tagsConfig.getString(e.getKey))).toMap
+          } else {
+            Map()
+          }
+        TagSet.from(tags)
+      }
       val nullValue: Option[String] =
         if (config.hasPath("null-value")) {
           Option(config.getString("null-value"))
         } else None
 
-      def generateSettingItem(_tags: kamon.Tags, tenant: Option[Tenant]) =
+      def generateSettingItem(_tags: kamon.tag.TagSet, tenant: Option[Tenant]) =
         SettingItem(key = MetricsKey(key, tenant), name = config.getString("name"), tags = _tags, nullValue = nullValue)
 
       val tenantSettingItems = tenants.map { implicit tenant =>
         import MetricsMultiTenantSupport._
-        generateSettingItem(tags.withTenant, Option(tenant))
+        generateSettingItem(tagSet.withTenant, Option(tenant))
       }
 
       // system-metrics など テナントに関係無いものとテナント別のものを別で管理する
-      tenantSettingItems + generateSettingItem(tags, tenant = None)
+      tenantSettingItems + generateSettingItem(tagSet, tenant = None)
     }
   }