+++ CHANGELOG.md
@@ -11,6 +11,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `lerna-testkit`: Added testkit for TypedActor
 - `lerna-log`: Added Logger for TypedActor
 
+### Changed
+- `lerna-management`
+  - Update to `Kamon 2.1.8` from `Kamon 1.1.6`
+  - Improve documentation
+  - Provide [migration guide](doc/migration-guide.md)
 
 ## [v1.0.0] - 2020-12-22
 [v1.0.0]: https://github.com/lerna-stack/lerna-app-library/tree/v1.0.0